ASNIDE-21: Renamed subdirectory from 'autocompletion' to 'completion'. Introduced new namespace 'Completion' for files from 'completion' subdirectory. Rearranged project files list in .pro file.
