+++ completion/acncompletionassist.cpp
@@ -23,32 +23,33 @@
 **
 ****************************************************************************/
 
-#pragma once
+#include "acncompletionassist.h"
 
-#include <coreplugin/id.h>
+#include "../asn1acnconstants.h"
 
-#include "completionassist.h"
+#include "acnbuiltinsproposalsprovider.h"
 
-namespace Asn1Acn {
-namespace Internal {
+using namespace Asn1Acn::Internal::Completion;
 
-class AsnCompletionAssistProcessor : public CompletionAssistProcessor
+AcnCompletionAssistProcessor::AcnCompletionAssistProcessor()
+    : CompletionAssistProcessor(QLatin1String(Constants::ACN1_SNIPPETS_GROUP_ID))
 {
-public:
-    AsnCompletionAssistProcessor();
+}
 
-private:
-    std::unique_ptr<BuiltinsProposalsProvider> getBuiltinsProposalsProvider() const override;
-};
-
-class AsnCompletionAssistProvider : public CompletionAssistProvider
+std::unique_ptr<BuiltinsProposalsProvider> AcnCompletionAssistProcessor::getBuiltinsProposalsProvider() const
 {
-    Q_OBJECT
+    auto provider = std::make_unique<AcnBuiltinsProposalProvider>();
+    return std::move(provider);
+}
 
-public:
-    bool supportsEditor(Core::Id editorId) const override;
-    TextEditor::IAssistProcessor *createProcessor() const override;
-};
+bool AcnCompletionAssistProvider::supportsEditor(Core::Id editorId) const
+{
+    return editorId == Constants::ACNEDITOR_ID;
+}
 
-} // Internal
-} // Asn1Acn
+TextEditor::IAssistProcessor *AcnCompletionAssistProvider::createProcessor() const
+{
+    return new AcnCompletionAssistProcessor;
+}
+
+
