+++ parseddocument.h
@@ -35,7 +35,7 @@
 
 #include "modeltreenode.h"
 #include "documentsourceinfo.h"
-#include "autocompletion/usertypesproposalsprovider.h"
+#include "completion/usertypesproposalsprovider.h"
 
 namespace Asn1Acn {
 namespace Internal {
@@ -53,7 +53,7 @@ public:
     Data::TypeReference getTypeReference(const int line, const int col) const;
     Data::SourceLocation getDefinitionLocation(const QString& typeAssignmentName, const QString& moduleName) const;
 
-    UserTypesProposalsProvider getProposalsProvider() const;
+    Completion::UserTypesProposalsProvider getProposalsProvider() const;
 
 private:
     ModelTreeNode::ModelTreeNodePtr createDefinition(const std::unique_ptr<Data::Definitions> &definition) const;
